End of generation
generation.pas is successfully ended. It sorts the Game (type Jeu) and the 2 initilialized player (type Joueur). In commun.pas, 'Array[1..NBOAT] of Bateau' was replaced by a specific type 'listeBateau' in order to avoid a bug due to the compiler. Also  adding the pdf with all the signatures and the schematic of the game.